Validação do delete

+++ delete.php
@@ -3,10 +3,14 @@
 <?php
 if (isset($_GET["codigo"])) {
 	$db = new SQLite3("pizzaria.db");
-	$db->exec("PRAGMA foreign_keys = ON");
-	$db->exec("delete from saboringrediente where sabor = " .$_GET["codigo"]);
-	$db->exec("delete from sabor where sabor.codigo = ".$_GET["codigo"]);
-	echo $db->changes()." Pizza(s) excluída(s)";
+	$results = $db->query("select * from pizzasabor where sabor= " . $_GET["codigo"]);
+	if($row = $results->fetchArray()) {
+		echo "<p style=\"color:red\"> O sabor foi utilizado em uma pizza, não pode ser eliminado!</p>";
+	} else {
+		$db->exec("delete from saboringrediente where sabor = " .$_GET["codigo"]);
+		$db->exec("delete from sabor where sabor.codigo = ".$_GET["codigo"]);
+		echo $db->changes()." Pizza(s) excluída(s)";
+	}
 	$db->close();
 }
 ?>
